Fix Disk Drive Labeling
ILO5 on Gen10 machines produce a different API output with regards to disk location.  This had the
effect of producing empty and therefore omitted location labels.  The flow on of that was that there
was no differentiating label between disks of the same model meaning disks were not represented in
the metrics.

This change adds support for the new location data structure as well as includes the disk Name and
serial number for extra differentiation.